// src/flux/type.ts
export enum Type {
	CREATE_COUNTER = 'CREATE_COUNTER',
	CREATE_PROJECT = 'CREATE_PROJECT',
	DELETE_COUNTER = 'DELETE_COUNTER',
	DELETE_PROJECT = 'DELETE_PROJECT',
	ENABLE_COUNTER = 'ENABLE_COUNTER',
	MANAGE_COUNTER = 'MANAGE_COUNTER',
	MANAGE_PROJECT = 'MANAGE_PROJECT',
	RESET_COUNTER = 'RESET_COUNTER',
	RESTORE_FROM_LOCAL_STORAGE = 'RESTORE_FROM_LOCAL_STORAGE',
	TOGGLE_THEME = 'TOGGLE_THEME',
	UNDO = 'UNDO',
	UPDATE_COUNTER = 'UPDATE_COUNTER'
}